requirements.txt
rich
pyfiglet
Pillow
